+++ src/Modules/Region/components/ThreadPost.vue
@@ -1,14 +1,14 @@
 <template>
-  <div :id="`post-${id}`" class="thread">
+  <div :id="`post-${post.id}`" class="thread">
     <div
       ref="card"
-      class="card mb-2"
-      :class="{'disabledLoading': isLoading}"
+      class="card mb-2 post-card"
+      :class="{'disabledLoading': isLoading, 'hidden': post.hidden}"
     >
       <div class="card-header d-flex align-items-center justify-content-between">
         <Avatar
-          v-if="wXS"
-          :user="author"
+          v-if="wXS || post.hidden"
+          :user="post.author"
           class="mr-2"
         />
         <span class="flex-grow-1">
@@ -17,24 +17,25 @@
             v-b-tooltip="$i18n('thread.post.linked_post')"
             class="fas fa-link mr-1"
           />
-          <a :href="$url('profile', author.id)">
-            <strong class="author">{{ author.name }}</strong>
-          </a>
+          <a :href="$url('profile', post.author.id)">
+            <strong class="author">{{ post.author.name }}</strong>
+          </a><!--
+          --><i v-if="post.hidden" v-text="': ' + $i18n('forum.post.hiddenPost', { reason: post.hidden })" />
         </span>
         <Time
-          :time="createdAt"
+          :time="post.createdAt"
           class="text-right"
         />
         <OverflowMenu :options="overflowMenuOptions" />
       </div>
-      <div class="d-flex m-2">
+      <div v-if="!post.hidden" class="d-flex m-2">
         <div
           v-if="!wXS"
           class="mr-2 pr-2 border-right border-light text-center"
           style="width: min-content; min-width: 150px;"
         >
           <Avatar
-            :user="author"
+            :user="post.author"
             class="mb-2"
             :size="130"
           />
@@ -48,22 +49,51 @@
           </a>
         </div>
         <div class="body m-2 mr-md-5 text-break">
-          <Markdown :source="body" />
+          <Markdown :source="post.body" />
         </div>
       </div>
-      <div class="card-footer text-right">
+      <div v-if="!post.hidden" class="card-footer text-right">
         <ThreadPostActions
-          :reactions="reactions"
-          :may-delete="mayDelete"
-          :may-edit="mayEdit"
+          :reactions="post.reactions"
+          :may-delete="post.mayDelete"
           :may-reply="mayReply"
+          :may-hide="mayHide"
           @delete="$emit('delete')"
+          @hide="$emit('hide', $event)"
           @reaction-add="$emit('reaction-add', $event)"
           @reaction-remove="$emit('reaction-remove', $event)"
-          @reply="$emit('reply', body)"
+          @reply="$emit('reply', post.body)"
         />
       </div>
     </div>
+    <b-modal
+      v-if="mayModerate"
+      ref="restoreModal"
+      :title="$i18n('forum.restore.title')"
+      centered
+      :cancel-title="$i18n('button.cancel')"
+      :ok-title="$i18n('button.yes_i_am_sure')"
+      @ok="restore"
+    >
+      {{ $i18n('forum.restore.really') }}
+      <blockquote>
+        <Markdown :source="post.body" />
+      </blockquote>
+      <ul v-if="hiddenDetails">
+        <li>
+          {{ $i18n('forum.restore.hidden_by') }}
+          <a :href="$url('profile', hiddenDetails.moderatorId)" v-text="hiddenDetails.moderatorName" />
+        </li>
+        <li v-text="$i18n('forum.restore.reason', { reason: post.hidden })" />
+        <li>
+          <Time
+            :time="hiddenDetails.time"
+            normal-size
+            :muted="false"
+          />
+        </li>
+      </ul>
+    </b-modal>
   </div>
 </template>
 
@@ -72,46 +102,71 @@ import Avatar from '@/components/Avatar/Avatar.vue'
 import ThreadPostActions from './ThreadPostActions'
 import conversationStore from '@/stores/conversations'
 import MediaQueryMixin from '@/mixins/MediaQueryMixin'
+import CopyToClipboardMixin from '@/mixins/CopyToClipboardMixin'
 import Markdown from '@/components/Markdown/Markdown.vue'
 import OverflowMenu from '@/components/OverflowMenu.vue'
 import Time from '@/components/Time.vue'
-import CopyToClipboardMixin from '@/mixins/CopyToClipboardMixin'
+import { pulseSuccess } from '@/script'
+import * as api from '@/api/forum'
 
 export default {
   components: { Avatar, ThreadPostActions, Markdown, OverflowMenu, Time },
   mixins: [MediaQueryMixin, CopyToClipboardMixin],
   props: {
-    id: { type: Number, default: null },
+    post: { type: Object, required: true },
     userId: { type: Number, required: true },
-    body: { type: String, default: '' },
-    author: { type: Object, default: () => ({ avatar: null }) },
-    createdAt: { type: Date, default: null },
     deepLink: { type: String, default: '' },
     reactions: { type: Object, default: () => ({}) },
-    mayEdit: { type: Boolean, default: false },
-    mayDelete: { type: Boolean, default: false },
+    mayHide: { type: Boolean, default: false },
+    mayModerate: { type: Boolean, default: false },
     isLoading: { type: Boolean, default: true },
     mayReply: { type: Boolean, default: true },
     isLinked: { type: Boolean, default: false },
   },
+  data: () => ({
+    hiddenDetails: null,
+  }),
   computed: {
     isMe () {
-      return this.userId === this.author.id
+      return this.userId === this.post.author.id
     },
     overflowMenuOptions () {
       return [
-        { hide: !navigator.clipboard, icon: 'copy', textKey: 'thread.post.options.copy_source', callback: () => this.copyToClipboard(this.body, 'thread.post.copy_source_success') },
+        { hide: !navigator.clipboard || this.post.hidden, icon: 'copy', textKey: 'thread.post.options.copy_source', callback: this.copySourceCodeToClipboard },
         { icon: 'chain', textKey: 'thread.post.options.copy_direct_link', callback: this.copyDirectLink },
+        { hide: !this.post.hidden || !this.mayModerate, icon: 'eye', textKey: 'thread.post.options.restore', callback: this.showRestoreModal },
       ]
     },
   },
   methods: {
     openChat () {
-      conversationStore.openChatWithUser(this.author.id)
+      conversationStore.openChatWithUser(this.post.author.id)
+    },
+    async copySourceCodeToClipboard () {
+      await navigator.clipboard.writeText(this.post.body)
+      pulseSuccess(this.$i18n('thread.post.copy_source_success'))
     },
     async copyDirectLink () {
       this.copyToClipboard(location.protocol + '//' + location.host + this.deepLink, 'thread.post.copy_direct_link_success')
     },
+    async restore () {
+      this.$emit('restore')
+    },
+    async showRestoreModal () {
+      if (!this.hiddenDetails) {
+        this.hiddenDetails = await api.getHiddenPostDetails(this.post.id)
+      }
+      this.$refs.restoreModal.show()
+    },
   },
 }
 </script>
+<style lang="scss" scoped>
+.post-card {
+  border: 1px solid var(--fs-border-default);
+}
+.hidden {
+  border: 1px solid var(--fs-color-danger-400);
+  background-color: var(--fs-color-danger-200);
+}
+</style>